test: fix unit tests
